use int64_t not ULL

+++ libomnireader/include/xdrthing.h
@@ -7,6 +7,7 @@
 
 #include <cstring>
 #include <string>
+#include <cstdint>
 
 class XDRImpl;
 
@@ -21,8 +22,8 @@ public:
     size_t get_double(const char *src, double &output) const;
     size_t get_int32(const char *src, int &output) const;
     size_t get_uint32(const char *src, unsigned int &output) const;
-    size_t get_int64(const char *src, long long &output) const;
-    size_t get_uint64(const char *src, unsigned long long &output) const;
+    size_t get_int64(const char *src, int64_t &output) const;
+    size_t get_uint64(const char *src, uint64_t &output) const;
     bool is_big_endian() const;
 private:
     XDRImpl* _impl;
